requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
streamlit>=1.39.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
